refactor: apply pagination to me page
