+++ client/src/pages/Me.tsx
@@ -2,15 +2,19 @@ import React, {ChangeEvent, useCallback, useEffect, useState} from "react";
 import {useRecoilValue, useResetRecoilState} from "recoil";
 import SidebarLayout from "../components/common/SidebarLayout";
 import Info from "../components/common/Info";
-import Cards from "../components/common/Cards";
 import usersApi from "../api/user";
 
-import {
-  getAllSlidesQuery,
-  userInfoQuery,
-} from "../store/atoms";
+import {userInfoQuery,} from "../store/atoms";
 import filesApi from "../api/file";
 import {googleAnalyticsEvent, googleAnalyticsPageView} from "../utils/googleAnalytics";
+import SlidesLayout from "../components/common/SlidesLayout";
+import styled from "@emotion/styled";
+
+const MeBlock = styled.div`
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+`
 
 export interface User {
   name: string,
@@ -19,7 +23,6 @@ export interface User {
 }
 
 const Me: React.FC = () => {
-  const slides = useRecoilValue(getAllSlidesQuery);
   const user = useRecoilValue(userInfoQuery);
   const setUser = useResetRecoilState(userInfoQuery);
   const [editedUser, setEditedUser] = useState<User>(user!);
@@ -52,12 +55,14 @@ const Me: React.FC = () => {
   return (
     <SidebarLayout>
       {/* <Cards title="Recent"/>*/}
-      <Info user={user!}
-        editedUser={editedUser}
-        updateInfo={updateInfo}
-        changeInput={changeInput}
-        changeProfileImage={changeProfileImage}/>
-      <Cards title="My Drafts" slides={slides} author={user?.name}/>
+      <MeBlock>
+        <Info user={user!}
+              editedUser={editedUser}
+              updateInfo={updateInfo}
+              changeInput={changeInput}
+              changeProfileImage={changeProfileImage}/>
+        <SlidesLayout slidesCnt={6} title="My Drafts"/>
+      </MeBlock>
     </SidebarLayout>
   );
 };
